requirements.txt
azureml-mlflow==1.56.0
azure-ai-ml==1.18.0
beautifulsoup4==4.12.3
executing==2.0.1
Faker==22.6.0
fastapi==0.109.2
gunicorn==21.2.0
hydra-core==1.3.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
mkdocs==1.5.3
mkdocs-autorefs==1.0.0
mkdocs-material==9.5.11
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.1
mkdocstrings-crystal==0.3.7
mkdocstrings-python==1.8.0
mlflow==2.12.1
mssql==1.0.1
mssql-django==1.5
numpy==1.26.4
omegaconf==2.3.0
openai==1.23.2
openpyxl==3.1.2
pandas==2.2.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
pyodbc==5.0.1
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-mock==3.14.0
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
SQLAlchemy==2.0.27
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.1
azure-monitor-opentelemetry-exporter==1.0.0b27
opentelemetry-exporter-otlp==1.25.0 
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-requests==0.46b0
sentence-transformers==3.0.0
pytest-django==4.8.0